move default link config to drop-in snippet
because otherwise we need to house *all* defaults, like the interface
naming policy ones, too.

This can be fine for one release, but easily overlooked if those, or
other important fall-back defaults change.

A user can now also easier override this, e.g., by simply adding a
drop-in file in the respective /etc path.

Fixes failure to rename network names to "predictable" names on boot
as reported in the forum:

https://forum.proxmox.com/threads/135635/page-6#post-606130

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>