ui: iso selector: disable all fields to avoid bogus validation
The validation logic of the inner fields from the ISO selector was not
disabled, so a user would need to select a valid storage and ISO file
before being able to make any other choice for the general CD-ROM
drive source (no-media or physical-drive).

Call the parent method to ensure all the inner fields get actually
disabled so that their validators also get disarmed if not relevant.

Reported in the communiy forum:
https://forum.proxmox.com/threads/136960/post-611704

Fixes: fc7b556d ("ui: refactor iso-selector out of the cd input panel")
Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>
 [ TL: add more background to commit message/subject ]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>