ui: esxi storage edit: make user and password always editable
Those can change somewhat frequently, e.g., as part of credentials
rotation policies.

Keep the 'server' one for now as is, while it can make sense to change
that too, e.g., from IP to FQDN, it's currently uncommon for PVE to
allow that in the UI, if we should evaluate all storage types for such
things and for that we also should improve the re-mount/activate story
in the storage lib.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>