ui: guest import: actually copy vmconfig state before applying values
Otherwise changes will persist, e.g., disabling a CDROM drive won't
work if the values got assembled for the "Resulting Config" tab
before. Same for the "Prepare for VirtIO" checkbox, there one got the
disk twice, once as (old) mapped SATA and once as (original) SCSI.

To fix this do not work on a reference of the original VM config, but
rather always copy that to a new object before applying the form
values.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>