ui: backup job: rework empty-text for advanced fields again
This partially reverts commit a32a5c4a6 ("ui: backup job: rework hint
about fallback config and make it less flashy"), i.e., the part about
the fallback values, as those was barely visible now.

Add the schema default to the end of the description and expand the
hint at the bottom to also mention that this is used as second level
fallback, if the vzdump.conf does not has the option set.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>