ui: backup job editor: fix disable behavior for fleecing storage
commit 569b0388 (ui: fix reset behavior of backup job editor) disabled
the fleecing storage field by default (as that is the default state)
to fix the reset behavior. This broke editing the job when fleecing
was enabled and the user did not navigate to the advanced tab yet.

It seems that the 'bind' here only gets triggered once the panel is
rendered, but we actually need it before that.

To work around the issue for now, manually enable/disable the field
when toggling the fleecing checkbox. (Though this warrants a bit of
deeper investigation into this bind behavior)

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>