ui: fix align mode of two column container
'stretch' is most often the wrong value, as that will stretch
everything, to the height of the whole container, including fields.
That is not desirable, since fields look not good when stretched this
way (e.g. the controls are not correctly aligned).

To fix it, simply set it to 'begin'.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>