www: backup: clarify experimental change detection modes
Currently, the whole mode selector is labeled as experimental, this
does however give the impression that also the default legacy mode is
an experimental mode.
To clarify that only the `data` and `metadata` change detection modes
are experimental, move the experimental label to the individual
modes and explicitly mention the experimental modes in the message.

Also, make it more clear that the archive encoding format depends on
the selected mode.

Signed-off-by: Christian Ebner <c.ebner@proxmox.com>